Refactor acme_protocol filename to reflect actual namespace
