+++ src/middlewared/middlewared/api/v25_04_0/__init__.py
@@ -1,5 +1,5 @@
-from .acme_protocol import *  # noqa
 from .acl import *  # noqa
+from .acme_registration import *  # noqa
 from .alert import *  # noqa
 from .alertservice import *  # noqa
 from .api_key import *  # noqa
